Make RTD reader constructor private, move config
In order to make this nested type opaque, use the same protocol + private
constructor subclass layout as the reader and writer. The config is moved
to be colocated with the relevant component.
